+++ iextrading4j-client/src/main/java/pl/zankowski/iextrading4j/client/IEXCloudClient.java
@@ -3,15 +3,14 @@ package pl.zankowski.iextrading4j.client;
 import pl.zankowski.iextrading4j.client.rest.manager.RestRequest;
 import pl.zankowski.iextrading4j.client.sse.manager.SseRequest;
 
-import java.io.Serializable;
 import java.util.function.Consumer;
 
 public interface IEXCloudClient {
 
     <R> R executeRequest(final RestRequest<R> restRequest);
 
-    <R extends Serializable> void subscribe(SseRequest<R> asyncRequest, Consumer<R> consumer);
+    <R> void subscribe(SseRequest<R> asyncRequest, Consumer<R> consumer);
 
-    <R extends Serializable> void unsubscribe(SseRequest<R> asyncRequest);
+    <R> void unsubscribe(SseRequest<R> asyncRequest);
 
 }