+++ iextrading4j-client/src/main/java/pl/zankowski/iextrading4j/client/rest/request/options/OptionSide.java
@@ -1,0 +1,18 @@
+package pl.zankowski.iextrading4j.client.rest.request.options;
+
+public enum OptionSide {
+
+    PUT("put"),
+    CALL("call");
+
+    private final String name;
+
+    OptionSide(final String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+}