+++ iextrading4j-client/src/test/java/pl/zankowski/iextrading4j/client/sse/request/marketdata/TopsSseRequestBuilderTest.java
@@ -6,6 +6,8 @@ import pl.zankowski.iextrading4j.client.sse.manager.SseRequest;
 
 import javax.ws.rs.core.GenericType;
 
+import java.util.List;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
 
@@ -15,12 +17,12 @@ public class TopsSseRequestBuilderTest {
     public void shouldSuccessfullyCreateSseRequest() {
         final String symbol = "IBM";
 
-        final SseRequest<TOPS> request = new TopsSseRequestBuilder()
+        final SseRequest<List<TOPS>> request = new TopsSseRequestBuilder()
                 .withSymbol(symbol)
                 .build();
 
         assertThat(request.getPath()).isEqualTo("/tops");
-        assertThat(request.getResponseType()).isEqualTo(new GenericType<TOPS>() {
+        assertThat(request.getResponseType()).isEqualTo(new GenericType<List<TOPS>>() {
         });
         assertThat(request.getPathParams()).isEmpty();
         assertThat(request.getQueryParams()).contains(entry("nosnapshot", "false"), entry("symbols", symbol));