+++ iextrading4j-samples/src/main/java/pl/zankowski/iextrading4j/sample/iexcloud/rest/RefDataSample.java
@@ -35,8 +35,7 @@ import pl.zankowski.iextrading4j.client.rest.request.refdata.v1.UsHolidayAndTrad
 import java.util.List;
 import java.util.Map;
 
-public class
-RefDataSample {
+public class RefDataSample {
 
     final IEXCloudClient cloudClient = IEXTradingClient.create(IEXTradingApiVersion.IEX_CLOUD_BETA_SANDBOX,
             new IEXCloudTokenBuilder()